Fix printf type/arg mismatches.

+++ tests/t-constants.c
@@ -210,7 +210,7 @@ main (int argc, char *argv[])
 {
   int  error = 0;
 
-  CHECK_INT (BYTES_PER_MP_LIMB, sizeof(mp_limb_t));
+  CHECK_INT (BYTES_PER_MP_LIMB, (int) sizeof(mp_limb_t));
   CHECK_INT (mp_bits_per_limb, BITS_PER_MP_LIMB);
   CHECK_INT (__GMP_BITS_PER_MP_LIMB, BITS_PER_MP_LIMB);
